requirements.txt
nfcpy
wrapt-timeout-decorator